@@ -4,7 +4,7 @@ version: 0.2.1
 homepage: https://www.github.com/brianegan/giphy_client
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   meta: ^1.1.2
